moved add resultstodisplay to utils directory
