+++ flipline.py
@@ -1,10 +1,9 @@
-import sys
 import os.path
+import sys
 from collections import OrderedDict
 
 import arcpy
-from addresulttodisplay import add_result_to_display
-
+from utils.addresulttodisplay import add_result_to_display
 
 sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), 'external'))
